requirements.txt
Flask
pillow
tensorflow
numpy
    